post

+++ app/Http/Controllers/TasksController.php
@@ -15,10 +15,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if(\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->get();
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks
+            ];
+        }
+        
+        return view('welcome', $data);
     }
 
     /**
@@ -28,10 +36,13 @@ class TasksController extends Controller
      */
     public function create()
     {
-        $task = new Task;
-        return view('tasks.create', [
-           'task' => $task,
-        ]);
+        if(\Auth::check()){
+            $task = new Task;
+            return view('tasks.create', [
+               'task' => $task,
+            ]);
+        }    
+        return view('welcome', $data);
     }
 
     /**
@@ -47,10 +58,10 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content
+        ]);
         
         return redirect('/');
     }
@@ -78,11 +89,13 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::findOrFail($id);
-        
-        return view('tasks.edit', [
-           'task' => $task,
-        ]);
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id)
+        {
+            return view('tasks.edit', [
+                'task' => $task
+            ]);
+        }
     }
 
     /**
@@ -99,12 +112,13 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        $task = Task::findOrFail($id);
-        
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id)
+        {
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
         return redirect('/');
     }
 
@@ -116,9 +130,10 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
-        $task->delete();
-        
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id){
+            $task->delete();
+        }
         return redirect('/');
     }
 }
